SVN r4269 (same idea)
(This commit in SVN changes from using sqrt to sqrtf and from using fabs
to using fabsf when assigning to a float value. Here this is just one
change, in a different file from SVN, from sqrt to sqrtf)

+++ include/util_math.h
@@ -27,7 +27,7 @@ struct DOSBox_Vector2
 	}
 
 	inline float magnitude(void) const {
-		return (float)sqrt(sqrMagnitude());
+		return sqrtf(sqrMagnitude());
 	}
 
 	inline float sqrMagnitude(void) const {
